+++ src/puppeteer.js
@@ -157,8 +157,7 @@ const getPuppeteerOrThrow = () => {
  *    <li>
  *        The function adds <code>--no-sandbox</code> to <code>args</code> to enable running
  *        headless Chrome in a Docker container on the Apify platform.
- *        Also it adds <code>--disable-background-timer-throttling</code>
- *        and <code>--enable-resource-load-scheduler=false</code>
+ *        Also it adds <code>--enable-resource-load-scheduler=false</code>
  *        to make crawling of pages in all tabs run equally fast.
  *    </li>
  * </ul>
@@ -197,7 +196,6 @@ export const launchPuppeteer = (options = {}) => {
 
     optsCopy.args = optsCopy.args || [];
     optsCopy.args.push('--no-sandbox');
-    optsCopy.args.push('-—disable-background-timer-throttling');
     // TODO: It's not clear that this works, keep eye on https://bugs.chromium.org/p/chromium/issues/detail?id=723233
     optsCopy.args.push('--enable-resource-load-scheduler=false');
     if (optsCopy.headless == null) {