+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+xxx
+==================
+- Automatically adding `--disable-background-timer-throttling` and `--enable-resource-load-scheduler=false`
+  Chrome flags in `Apify.launchPuppeteer()`
+  to make crawling of pages in all tabs run equally fast.
+
 0.9.3 / 2018-11-12
 ==================
 - Bug fixes and improvements of internals.